pyproject.toml
[tool.poetry]
name = "jaeger"
version = "1.7.0a0"
description = "Controllers for the SDSS-V FPS"
authors = ["José Sánchez-Gallego <gallegoj@uw.edu>"]
license = "BSD-3-Clause"
readme = "README.md"
homepage = "https://github.com/sdss/jaeger"
repository = "https://github.com/sdss/jaeger"
documentation = "https://sdss-jaeger.readthedocs.io/en/latest/"
keywords = ["astronomy", "software"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Topic :: Documentation :: Sphinx",
    "Topic :: Software Development :: Libraries :: Python Modules"
]
packages = [
    { include = "jaeger", from = "src" }
]
include = ["src/jaeger/etc/*"]

[tool.poetry.dependencies]
python = "^3.10,<3.13"
sdss-drift = "^1.0.0"
sdss-clu = "^2.0.0"
sdsstools = "^1.6.0"
sdssdb = "^0.11.3"
sdss-kaiju =  [
    {version="^1.4.0b1", python=">=3.11"},
    {version="^1.3.0,<1.4.0", python=">=3.10,<3.11"}
]
sdss-coordio = "^1.9.2"
click_default_group = ">=1.2.2"
"zc.lockfile" = "^2.0"
polars = "^0.20.9"
adbc-driver-postgresql = ">=0.9.0"
pyarrow = ">=15.0.0"
tables = "^3.9.2"
nptyping = "^2.5.0"

[tool.poetry.group.dev.dependencies]
ipython = ">=8.0.0"
invoke = ">=1.3.0"
flake8 = ">=3.7.9"
doc8 = ">=0.8.0"
pytest = ">=5.2.2"
progressbar2 = ">=3.39.3"
pytest-asyncio = ">=0.10.0"
pytest-cov = ">=2.8.1"
pytest-sugar = ">=0.9.2"
isort = ">=5.0.0"
coverage = {version = ">=5.0", extras = ["toml"]}
Sphinx = ">=7.0"
sphinx-click = ">=2.3.0"
ipdb = ">=0.13.3"
autopep8 = ">=1.5.4"
black = ">=24.4.2"
rstcheck = ">=3.3.1"
pytest-mock = ">=3.5.1"
pyserial-asyncio = ">=0.5"
sphinx-jsonschema = ">=1.16.8"
myst-parser = ">=0.14.0"
furo = ">=2021.6.18-beta.36"
nox = ">=2021.6.12"
sphinx-autobuild = ">=2021.3.14"
sphinx-copybutton = ">=0.3.3"
matplotlib = ">=3.4.3"
ruff = ">=0.1.0"

[tool.poetry.scripts]
jaeger = "jaeger.__main__:jaeger"
robotcalib = "jaeger.scripts.robotcalib:robotcalib"

[tool.black]
line-length = 88
target-version = ['py312']
fast = true

[tool.ruff]
line-length = 88
target-version = 'py312'
exclude = ["typings/"]

[tool.ruff.lint]
select = ["E", "F", "I"]
unfixable = ["F841"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "F403", "E402"]

[tool.ruff.lint.isort]
known-first-party = ["jaeger"]
lines-after-imports = 2
section-order = ["future", "standard-library", "typing", "third-party", "sdss", "first-party", "local-folder"]

[tool.ruff.lint.isort.sections]
typing = ["typing"]
sdss = ["sdsstools", "clu", "sdssdb", "drift", "coordio", "kaiju", "fps_calibrations"]

[tool.pytest.ini_options]
addopts = "-p no:logging --disable-warnings --cov jaeger --cov-report xml --cov-report html --cov-report term"
asyncio_mode = "auto"
filterwarnings = [
    "ignore:.*The loop argument is deprecated since.*",
    "ignore:.*Loading an empty FPS.*",
    "ignore:.*They have been added to the.*",
]
markers = [
    "rtd2: Temperature of the RTD2 sensor"
]

[tool.coverage.run]
branch = true
include = ["src/jaeger/*"]
omit = [
    "*/__main__.py",
    "*/__init__.py",
    "*/testing.py",
    "*/cannet.py",
    "*/interfaces/message.py",
    "*/actor/commands/bootloader.py",
    "*/actor/commands/talk.py",
    "*/commands/calibration.py",
    "*/exceptions.py",
    "*/scripts/*",
    "*/plotting.py"
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "class AsyncioExecutor",
    "class CANnetInterface",
    "if TYPE_CHECKING:",
    "def __repr__"
]

[build-system]
requires = ["poetry-core>=1.1.0", "setuptools"]
build-backend = "poetry.core.masonry.api"
